Added the useportal.euo script and corresponding example.
useportal.euo is utility script for porting via teleport system on
Andaria. It is used to automaticaly select and go to location specified by
the index of button on portal selection panel.

Example script porting.euo is designed to show it's functionality. You have to
stand into the middle of the main porting pad in Thyris and run the
script. The controll is returned once the character reaches the
"Islands" destination.