Added fishing.euo and twoandground.euo scripts.
sources/fishing.euo is (as the name suggests) script to fish.
Script sources/twoandground.euo is it's dependency. Simply
stand near some fishable watter and run the script. You will
get prompted whether or not to reset the fished out squares,
this is because after successful run of the script all
positions are memorized via win registery. Reseting is good
idea on dayly basis, so that the fish has some time to
respawn. At the end of the script you are informed so, and
the script is paused. You should go somwhere else and start
it over. This is the version of the script I am using, with
some slight modifications in order to make it not so
efficient.

WARNING: Make sure to have status and paperdoll open.

There is no example of usage of this script, if you are dumb
enough not to figure it out on your own, you are not allowed
to use it (YES this is my new EULA concerning this script).

Updated the example/clicking.euo to show the usage of
twoandground.euo script.